# vop-sales-worksheet/src/main/resources/openapi/openapi.yml
# https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.1.md
openapi: 3.0.1
info:
  version: 0.0.1-SNAPSHOT
  title: VOP Worksheet Service v1 API Documentation
  description: An OpenAPI 3.0 specification for Worksheet Service. Documentation is generated using OpenAPI Specification version 3.0.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
tags:
  - name: Worksheet
    description: Worksheet Resource
  - name: token-resource
    description: Token Resource
  - name: Tours
    description: Tour Resource
  - name: Package
    description: Package Resource
security:
  - bearerAuth: [ ]
paths:
  "/api/v1/worksheet/pid":
    post:
      tags:
        - Worksheet
      summary: Retrieve Worksheet information by Participant ID (PID) from mocked partner sample service.
      description: Will return a sample info object based on search by PID.
      operationId: sampleByPid
      requestBody:
        description: sampleRequest
        required: true
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/SampleRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/SampleResponse'
        '400':
          '$ref': '#/components/responses/400'
        '401':
          '$ref': '#/components/responses/401'
        '403':
          '$ref': '#/components/responses/403'
        default:
          '$ref': '#/components/responses/500'
        '500':
          '$ref': '#/components/responses/500'
      security:
        - bearerAuth: [ ]
      deprecated: false
  /api/v1/worksheet/packages/preset/{siteId}/{entityId}:
    get:
      tags:
        - Package
      summary: Retrieve preset packages information by site ID and entity ID.
      description: Returns preset packages information based on the provided site ID and entity ID.
      operationId: presetPackagesBySiteIdAndEntityId
      parameters:
        - name: siteId
          in: path
          required: true
          schema:
            type: string
          description: The site ID for which preset packages are retrieved.
        - name: entityId
          in: path
          required: true
          schema:
            type: string
          description: The entity ID for which preset packages are retrieved.
      responses:
        '200':
          description: Successfully retrieved preset packages information.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PresetPackageResponse'
              examples:
                example-1:
                  value:
                    salesPresetPackageInventoryId: 12345
                    presetPackageNumberOfCriteria: 10
                    presetPackageCreateOrUpdateDate: "2023-09-03T10:15:30"
                    presetPackageUpdateFlag: "UPDATED"
                    serviceEntity: "SERVICE_1"
                    siteId: 1001
                    inventoryType: "TYPE_A"
                    totalPoints: 200
                    developerPrice: 15000
                    defaultMinimumDownPercentage: 10.5
                    customerMinimumDownPercentage: 20.5
                    defaultInterestRate: 3.5
                    maxDiscountType: 1
                    customerMaxDiscount: 1000
                    maintenanceFee: 200
                    inventoryId: "INV001"
                    loanTerm: 60
                    customerLoanTerm: 60
                    processingFee: 300
                    bonusPointType: "BONUS_1"
                    active: true
                    createDate: "2023-08-01T12:00:00"
                    createUserId: "USER123"
                    modifyDate: "2023-08-05T12:00:00"
                    modifyUserId: "USER456"

        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        default:
          '$ref': '#/components/responses/500'
        '500':
          $ref: '#/components/responses/500'
      security:
        - bearerAuth: []
      deprecated: false
  # The token path should not require any changes
components:
  # securitySchemes are defined by the framework.
  # Service applications must configure as below.
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
  schemas:
    # Define API (provider) model objects. Use type, format and example;
    # use standard JSR303 constraints (required, min, max, etc) where it makes sense
    SampleInfo:
      type: object
      title: SampleInfo
      description: Model for sample data contained in the response from the Service
      properties:
        name:
          type: string
          example: JANE M DOE
          description: The worksheets first name, middle initial, and last name
        participantId:
          type: integer
          format: int64
          example: 6666345
          minimum: 1
          description: The worksheets participant ID
    SampleRequest:
      title: SampleRequest
      description: Model for data to request sample information from the Service
      allOf:
        - '$ref': '#/components/schemas/ProviderRequest'
        - type: object
          properties:
            participantID:
              type: integer
              format: int64
              example: 6666345
              description: The Participant ID of the worksheet for whom to retrieve sample data
              minimum: 1
              exclusiveMinimum: false
          required:
            - participantID
    SampleResponse:
      title: SampleResponse
      description: Model for the sample response from the Service
      allOf:
        - '$ref': '#/components/schemas/ProviderResponse'
        - type: object
          # all other properties specific to SampleResponse
          properties:
            sampleInfo:
              description: The object representing the sample information
              '$ref': '#/components/schemas/SampleInfo'
    PresetPackageResponse:
      title: PresetPackageResponse
      type: object
    # Schema objects below are from VOP Framework to be declared.
    # These objects need to be mapped in <importMappings> section
    # of openapi-generator-maven-plugin so that no code is generated
    # for these classes
    Message:
      type: object
      title: Message
      description: Model that identifies a single message added to the repsonse. Property keys are enumerated in  the MessageKeys object.
      required:
        - key
        - severity
      properties:
        key:
          # possible values enumerated by com.wynd.vop.framework.messages.MessageKeys
          type: string
        severity:
          # derived from com.wynd.vop.framework.messages.MessageSeverity
          type: string
          enum:
            - INFO
            - WARN
            - ERROR
        status:
          type: integer
          format: int32
        text:
          type: string
        timestamp:
          type: string
          format: date-time
          example: yyyy-MM-dd'T'HH:mm:ss.SSS
    Worksheet:
      type: object
      title: Worksheet
      description: Model that identifies a single individual used in the security context
      properties:
        applicationID:
          type: string
          example: WorksheetUI
          description: The application taking the last action on the record
        appToken:
          type: string
          example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9
          description: The application's custom token passed as any String from consumers to the API services
        birthDate:
          type: string
          format: date
          example: '1978-05-20'
          description: The worksheet's birth date
        email:
          type: string
          example: jane.doe@wyn.com
          description: The worksheet's email address
        firstName:
          type: string
          example: JANE
          description: The worksheet's first name
        gender:
          type: string
          example: FEMALE
          description: The worksheet's gender
        lastName:
          type: string
          example: DOE
          description: The worksheet's last name
        middleName:
          type: string
          example: M
          description: The worksheet's middle name
        prefix:
          type: string
          example: Ms
          description: The prefix for the worksheet's full name
        suffix:
          type: string
          example: S
          description: The suffix for the worksheet's full name
        userID:
          type: string
          example: vhaislXXXXX
          description: The name associated with the worksheet taking the last action on the record
    ProviderResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            '$ref': '#/components/schemas/Message'
      title: ProviderResponse
    ProviderRequest:
      type: object
      title: ProviderRequest
  responses:
    400:
      description: There was an error encountered processing the Request.  Response
        will contain a  "messages" element that will provide further information
        on the error.  This request shouldn't be retried until corrected.
      content:
        application/problem+json:
          schema:
            '$ref': '#/components/schemas/ProviderResponse'
          example:
            messages:
              - key: vop.framework.global.validator.method.argument.not.valid
                severity: ERROR
                status: 400
                text: Argument not valid.
                timestamp: 2019-08-29T18:40:22.766Z
    401:
      description: The authentication mechanism failed and hence access is forbidden.
      content:
        application/problem+json:
          schema:
            '$ref': '#/components/schemas/ProviderResponse'
          example:
            messages:
              - key: UNAUTHORIZED
                severity: ERROR
                status: 401
                text: No JWT Token in Header.
                timestamp: 2019-08-29T18:40:22.766Z
    403:
      description: The request is not authorized.  Please verify credentials used
        in the request.
      content:
        application/problem+json:
          schema:
            '$ref': '#/components/schemas/ProviderResponse'
          example:
            messages:
              - key: vop.framework.not.authorized.exception
                severity: ERROR
                status: 403
                text: Access denied.
                timestamp: 2019-08-29T18:40:22.766Z
    500:
      description: There was an error encountered processing the Request.  Response
        will contain a  "messages" element that will provide further information
        on the error.  Please retry.  If problem persists, please contact support
        with a copy of the Response.
      content:
        application/problem+json:
          schema:
            '$ref': '#/components/schemas/ProviderResponse'
          example:
            messages:
              - key: vop.framework.global.general.exception
                severity: ERROR
                status: 500
                text: Unexpected exception.
                timestamp: 2019-08-29T18:40:22.766Z